// packages/sheets-chart-ui/src/views/edit-chart/widget/interface.ts
/**
 * Copyright 2023-present DreamNum Co., Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface ISettingConfig {
    refString?: string;
    categoryIndex?: string | number;
    seriesIndexes?: (string | number)[];
    // eslint-disable-next-line ts/no-explicit-any
    [key: string]: any;
}

export interface ISelectOption {
    label: string;
    value: string;
}